+++ vector/src/main/java/im/vector/activity/CommonActivityUtils.java
@@ -623,7 +623,7 @@ public class CommonActivityUtils {
                 }
             }
 
-            if (null != EventStreamService.getInstance()) {
+            if (EventStreamService.getInstance() != null) {
                 EventStreamService.getInstance().refreshForegroundNotification();
             }
         }
@@ -1070,26 +1070,29 @@ public class CommonActivityUtils {
                                    final String roomId,
                                    final RoomPreviewData roomPreviewData,
                                    final ApiCallback<Void> callback) {
-        Room room = session.getDataHandler().getRoom(roomId, false);
+        // Check whether the room exists to handled the cases where the user is invited or he has joined.
+        // CAUTION: the room may exist whereas the user membership is neither invited nor joined.
+        final Room room = session.getDataHandler().getRoom(roomId, false);
+        if (null != room && room.hasMembership(RoomMember.MEMBERSHIP_INVITE)) {
+            Log.d(LOG_TAG, "previewRoom : the user is invited -> display the preview " + VectorApp.getCurrentActivity());
+            previewRoom(fromActivity, roomPreviewData);
 
-        // if the room exists
-        if (null != room) {
-            // either the user is invited
-            if (room.isInvited()) {
-                Log.d(LOG_TAG, "previewRoom : the user is invited -> display the preview " + VectorApp.getCurrentActivity());
-                previewRoom(fromActivity, roomPreviewData);
-            } else {
-                Log.d(LOG_TAG, "previewRoom : open the room");
-                Map<String, Object> params = new HashMap<>();
-                params.put(VectorRoomActivity.EXTRA_MATRIX_ID, session.getMyUserId());
-                params.put(VectorRoomActivity.EXTRA_ROOM_ID, roomId);
-                CommonActivityUtils.goToRoomPage(fromActivity, session, params);
+            if (null != callback) {
+                callback.onSuccess(null);
             }
+        } else if (null != room && room.hasMembership(RoomMember.MEMBERSHIP_JOIN)) {
+            Log.d(LOG_TAG, "previewRoom : the user joined the room -> open the room");
+            final Map<String, Object> params = new HashMap<>();
+            params.put(VectorRoomActivity.EXTRA_MATRIX_ID, session.getMyUserId());
+            params.put(VectorRoomActivity.EXTRA_ROOM_ID, roomId);
+            CommonActivityUtils.goToRoomPage(fromActivity, session, params);
 
             if (null != callback) {
                 callback.onSuccess(null);
             }
         } else {
+            // Display a preview by default.
+            Log.d(LOG_TAG, "previewRoom : display the preview");
             roomPreviewData.fetchPreviewData(new ApiCallback<Void>() {
                 private void onDone() {
                     if (null != callback) {