+++ vector/src/main/java/im/vector/fragments/VectorSettingsPreferencesFragment.kt
@@ -449,20 +449,11 @@ class VectorSettingsPreferencesFragment : PreferenceFragment(), SharedPreference
 
                     Matrix.getInstance(activity)!!.sharedGCMRegistrationManager
                             .forceSessionsRegistration(object : GcmRegistrationManager.ThirdPartyRegistrationListener {
-
-                                override fun onThirdPartyRegistered() {
+                                override fun onSuccess() {
                                     hideLoadingView()
                                 }
 
-                                override fun onThirdPartyRegistrationFailed() {
-                                    hideLoadingView()
-                                }
-
-                                override fun onThirdPartyUnregistered() {
-                                    hideLoadingView()
-                                }
-
-                                override fun onThirdPartyUnregistrationFailed() {
+                                override fun onError() {
                                     hideLoadingView()
                                 }
                             })
@@ -1102,7 +1093,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragment(), SharedPreference
 
             // when using GCM
             // need to register on servers
-            if (isConnected && gcmMgr.useGCM() && (gcmMgr.isServerRegistred || gcmMgr.isServerUnRegistred)) {
+            if (isConnected && gcmMgr.useGCM() && (gcmMgr.isServerRegistered || gcmMgr.isServerUnRegistered)) {
                 val listener = object : GcmRegistrationManager.ThirdPartyRegistrationListener {
 
                     private fun onDone() {
@@ -1114,27 +1105,19 @@ class VectorSettingsPreferencesFragment : PreferenceFragment(), SharedPreference
                         }
                     }
 
-                    override fun onThirdPartyRegistered() {
+                    override fun onSuccess() {
                         onDone()
                     }
 
-                    override fun onThirdPartyRegistrationFailed() {
-                        gcmMgr.setDeviceNotificationsAllowed(isAllowed)
-                        onDone()
-                    }
-
-                    override fun onThirdPartyUnregistered() {
-                        onDone()
-                    }
-
-                    override fun onThirdPartyUnregistrationFailed() {
+                    override fun onError() {
+                        // Set again the previous state
                         gcmMgr.setDeviceNotificationsAllowed(isAllowed)
                         onDone()
                     }
                 }
 
                 displayLoadingView()
-                if (gcmMgr.isServerRegistred) {
+                if (gcmMgr.isServerRegistered) {
                     gcmMgr.unregister(listener)
                 } else {
                     gcmMgr.register(listener)